refactor: extract log printing
