+++ cargo-shuttle/src/client.rs
@@ -4,9 +4,6 @@ use std::io::Read;
 use std::time::Duration;
 
 use anyhow::{anyhow, Context, Result};
-use chrono::{DateTime, Local};
-use colored::{ColoredString, Colorize};
-use log::Level;
 use reqwest::{Response, StatusCode};
 use reqwest_middleware::{ClientBuilder, ClientWithMiddleware};
 use reqwest_retry::policies::ExponentialBackoff;
@@ -14,6 +11,8 @@ use reqwest_retry::RetryTransientMiddleware;
 use shuttle_common::project::ProjectName;
 use shuttle_common::{ApiKey, ApiUrl, DeploymentMeta, DeploymentStateMeta, SHUTTLE_PROJECT_HEADER};
 use tokio::time::sleep;
+
+use crate::print;
 
 pub(crate) async fn auth(api_url: ApiUrl, username: String) -> Result<ApiKey> {
     let client = get_retry_client();
@@ -75,30 +74,11 @@ pub(crate) async fn logs(api_url: ApiUrl, api_key: &ApiKey, project: &ProjectNam
 
     let deployment_meta = get_deployment_meta(api_url, api_key, project, &client).await?;
 
-    for (datetime, log) in deployment_meta.runtime_logs {
-        let datetime: DateTime<Local> = DateTime::from(datetime);
-        println!(
-            "{}{} {:<5} {}{} {}",
-            "[".bright_black(),
-            datetime.format("%Y-%m-%dT%H:%M:%SZ"),
-            get_colored_level(&log.level),
-            log.target,
-            "]".bright_black(),
-            log.body
-        );
+    for (datetime, log_item) in deployment_meta.runtime_logs {
+        print::log(datetime, log_item);
     }
 
     Ok(())
-}
-
-fn get_colored_level(level: &Level) -> ColoredString {
-    match level {
-        Level::Trace => level.to_string().bright_black(),
-        Level::Debug => level.to_string().blue(),
-        Level::Info => level.to_string().green(),
-        Level::Warn => level.to_string().yellow(),
-        Level::Error => level.to_string().red(),
-    }
 }
 
 async fn get_deployment_meta(
